started zad 4 (BMI, BMR, home and recipes work)

+++ zad3/src/main/java/com/example/zad3/ui/game/Cannon.java
@@ -63,7 +63,7 @@ public class Cannon {
                 barrelEnd.y, paint);
 
         // draw cannon base
-        canvas.drawCircle(0, (int) view.getScreenHeight() / 2,
+        canvas.drawCircle(0,  view.getScreenHeight() / 2,
                 (int) baseRadius, paint);
     }
